PYIC-848: Build Docker image on merge to main
This adds a Github workflow to build a docker image for the smoke tests
and push it to Github.

The base image is a java image rather than a gradle image. This is so
it's easier to upgrade gradle without having to upgrade the image.
Gradle wrapper will actually install gradle for us.

This also commits the gradle-wrapper.jar. This is important so that
gradle wrapper can run when the image is being built. Without it,
gradle wrapper won't run.

Co-authored-by: Dan Worth <dan.worth@digital.cabinet-office.gov.uk>